split into two functions
