+++ fractal-terrain-generation/binaris.yml
@@ -3,3 +3,7 @@ functions:
     file: fractal.js
     entrypoint: handler
     runtime: node8
+  public_servePage:
+    file: servePage.js
+    entrypoint: handler
+    runtime: node8
